+++ tests/irdl/test_declarative_assembly_format.py
@@ -2119,51 +2119,3 @@ def test_default_properties(program: str, output: str, generic: str):
     printer.print_op(parsed)
 
     assert generic == stream.getvalue()
-
-
-def test_default_accessors():
-    ctx = MLContext()
-    ctx.load_op(DefaultOp)
-
-    parsed = Parser(ctx, "test.default").parse_operation()
-
-    assert isinstance(parsed, DefaultOp)
-
-    assert parsed.prop.value.data == 0
-
-    assert parsed.properties.get("opt_prop") is None
-
-    assert parsed.opt_prop.value.data == 1
-
-    assert parsed.attr.value.data == 0
-
-    assert parsed.attributes.get("opt_attr") is None
-
-    assert parsed.opt_attr.value.data == 1
-
-
-def test_generic_accessors():
-    ctx = MLContext()
-    ctx.load_op(DefaultOp)
-
-    parsed = Parser(
-        ctx, '"test.default"() <{ "prop" = false }> {"attr" = false} : () -> ()'
-    ).parse_operation()
-
-    assert isinstance(parsed, DefaultOp)
-
-    assert parsed.prop.value.data == 0
-
-    assert parsed.properties.get("opt_prop") is None
-
-    assert parsed.opt_prop is not None
-
-    assert parsed.opt_prop.value.data == 1
-
-    assert parsed.attr.value.data == 0
-
-    assert parsed.attributes.get("opt_attr") is None
-
-    assert parsed.opt_attr is not None
-
-    assert parsed.opt_attr.value.data == 1
